[fix][#36] organize 'test' packages
