+++ core/src/test/java/si/jernej/mexplorer/core/test/processing/TargetExtractionTest.java
@@ -1,0 +1,5 @@
+package si.jernej.mexplorer.core.test.processing;
+
+public class TargetExtractionTest
+{
+}
