+++ core/src/test/java/si/jernej/mexplorer/core/test/processing/transform/ValueTransformerTest.java
@@ -1,4 +1,4 @@
-package si.jernej.mexplorer.core.test.transform;
+package si.jernej.mexplorer.core.test.processing.transform;
 
 import static si.jernej.mexplorer.processorapi.v1.model.TransformDto.DateDiffRoundTypeEnum.FIFTEEN_YEARS;
 import static si.jernej.mexplorer.processorapi.v1.model.TransformDto.DateDiffRoundTypeEnum.FIVE_YEARS;
